pyproject.toml
[tool.poetry]
name = "iac-analysis"
version = "0.1.0"
description = ""
authors = ["Feitong Leo Qiao <flq2101@columbia.edu>"]
readme = "README.md"

[tool.poetry.scripts]
iac-analysis = "iac_analysis.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = { extras = ["all"], version = "^0.9.0" }
z3-solver = "^4.12.2.0"
pyyaml = "^6.0.1"
setuptools = "^69.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
